Added the currentuser auth on a separate file
